--- loans/document/serializers.py
@@ -1,9 +1,0 @@
-from rest_framework import serializers
-from .models import  Document
-
-class DocumentSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Document
-        fields = '__all__'
-
-
