--- loans/farmerLoan/serializers.py
@@ -1,7 +1,0 @@
-from rest_framework import serializers
-from .models import  Loan 
-
-class LoanSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Loan
-        fields = '__all__'
