--- loans/farmerLoan/urls.py
@@ -1,12 +1,0 @@
-
-
-from django.urls import path, include
-from rest_framework.routers import DefaultRouter
-from .views import LoanViewSet
-
-router = DefaultRouter()
-router.register(r'loans', LoanViewSet, basename="loans")
-
-urlpatterns = [
-    path('', include(router.urls)),
-]
