--- loans/farmerLoan/views.py
@@ -1,8 +1,0 @@
-
-from rest_framework import viewsets
-from .models import Loan
-from .serializers import LoanSerializer
-
-class LoanViewSet(viewsets.ModelViewSet):
-    queryset = Loan.objects.all()
-    serializer_class=LoanSerializer
