+++ loans/loans/urls.py
@@ -22,14 +22,14 @@ from django.urls import path,include
 urlpatterns = [
     path('admin/', admin.site.urls),
 
-    path('',include("loan_repayments.urls")),
+    # path('',include("loan_repayments.urls")),
 
-    path('',include("bankpartners.urls")),
-    path('farmerLoan/',include('farmerLoan.urls')),
-    path('', include('document.urls')),
-    path('',include("cooperative.urls")),
-    path('',include("farmer.urls")),
-    path('', include("farmerWealth.urls")),
+    # path('',include("bankpartners.urls")),
+    # path('farmerLoan/',include('farmerLoan.urls')),
+    path('api/', include('api.urls')),
+    # path('',include("cooperative.urls")),
+    # path('',include("farmer.urls")),
+    # path('', include("farmerWealth.urls")),
 ]
