+++ loans/requirements.txt
@@ -1,4 +1,3 @@
 asgiref==3.8.1
 django==5.2.3
-djangorestframework==3.16.0
 sqlparse==0.5.3
